banking_lms/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
python-multipart
python-dotenv
passlib[bcrypt]
